adding 2 func to itsonans (#2133)
* Implement it_802CD4FC and it_802CD7D4 functions

* Change x4 type from u32 to f32 in itsonansAttributes

* Add ItemStateTable for itSonans

* Clean up whitespace in itsonans.c

Removed unnecessary blank line in the item processing logic.

* Add include for itcoll.h in itsonans.c

* fix itsonansAttributes

* Add new constants for item properties

* Fix typo

+++ src/melee/it/itCommonItems.h
@@ -983,7 +983,7 @@ typedef struct itSScopeAttributes {
 
 typedef struct itsonansAttributes{
     f32 x0;
-    u32 x4;
+    f32 x4;
     f32 x8; //hit sens
     f32 xC;
     f32 x10;
